Fix mixed up storybook file names
